+++ src/OpenCensus/Resources/Resource.cs
@@ -14,13 +14,13 @@
 // limitations under the License.
 // </copyright>
 
-namespace OpenCensus.Resource
+namespace OpenCensus.Resources
 {
     using System;
     using System.Collections.Generic;
-    using System.Linq;
     using System.Security;
     using System.Text.RegularExpressions;
+    using OpenCensus.Implementation;
     using OpenCensus.Tags;
     using OpenCensus.Utils;
 
@@ -32,40 +32,14 @@ namespace OpenCensus.Resource
     public abstract class Resource : IResource
     {
         /// <summary>
-        /// Maximum length of the resource type name.
-        /// </summary>
-        internal const int MaxResourceTypeNameLength = 255;
-
-        /// <summary>
-        /// Special resource type name that is assigned if nothing else is detected.
-        /// </summary>
-        private const string GlobalResourceType = "Global";
-
-        /// <summary>
-        /// OpenCensus Resource Type Environment Variable Name
-        /// </summary>
-        private const string OC_RESOURCE_TYPE_ENV = "OC_RESOURCE_TYPE";
-
-        /// <summary>
-        /// OpenCensus Resource Labels Environment Variable Name
-        /// </summary>
-        private const string OC_RESOURCE_LABELS_ENV = "OC_RESOURCE_LABELS";
-
-        /// <summary>
         /// Tag list splitter.
         /// </summary>
-        private const char LABEL_LIST_SPLITTER = ',';
+        private const char LabelListSplitter = ',';
 
         /// <summary>
         /// Key-value splitter.
         /// </summary>
-        private const char LABEL_KEY_VALUE_SPLITTER = '=';
-
-        /// <summary>
-        /// Error message when string contains invalid characters.
-        /// </summary>
-        private static readonly string ERROR_MESSAGE_INVALID_CHARS =
-            " should be a ASCII string with a length greater than 0 and not exceed " + MaxResourceTypeNameLength + " characters.";
+        private const char LabelKeyValueSplitter = '=';
 
         /// <summary>
         /// Environment identification (for example, AKS/GKE/etc).
@@ -81,32 +55,32 @@ namespace OpenCensus.Resource
 
             try
             {
-                openCensusResourceType = Environment.GetEnvironmentVariable(OC_RESOURCE_TYPE_ENV);
+                openCensusResourceType = Environment.GetEnvironmentVariable(Constants.ResourceTypeEnvironmentVariable);
             }
-            catch (SecurityException)
+            catch (SecurityException ex)
             {
-                openCensusResourceType = GlobalResourceType;
+                openCensusResourceType = Constants.GlobalResourceType;
 
-                // TODO - Log error
+                Log.SecurityExceptionWarning(ex);
             }
 
             try
             {
-                openCensusEnvironmentTags = Environment.GetEnvironmentVariable(OC_RESOURCE_LABELS_ENV);
+                openCensusEnvironmentTags = Environment.GetEnvironmentVariable(Constants.ResourceLabelsEnvironmentVariable);
             }
-            catch (SecurityException)
+            catch (SecurityException ex)
             {
                 openCensusEnvironmentTags = string.Empty;
 
-                // TODO - Log error
+                Log.SecurityExceptionWarning(ex);
             }
 
             TryParseResourceType(openCensusResourceType, out EnvironmentType);
-            EnvironmentToLabelMap = ParseResourceLabels(Environment.GetEnvironmentVariable(OC_RESOURCE_LABELS_ENV));
+            EnvironmentToLabelMap = ParseResourceLabels(Environment.GetEnvironmentVariable(Constants.ResourceLabelsEnvironmentVariable));
         }
 
         /// <summary>
-        /// Gets the identification of the resource.
+        /// Gets or sets the identification of the resource.
         /// </summary>
         public abstract string Type { get; protected set; }
 
@@ -114,6 +88,8 @@ namespace OpenCensus.Resource
         /// Gets the map between the tag and its value.
         /// </summary>
         public abstract IList<ITag> Tags { get; }
+
+        private static OpenCensusEventSource Log => OpenCensusEventSource.Log;
 
         /// <summary>
         /// Creates a label/tag map from the OC_RESOURCE_LABELS environment variable.
@@ -123,8 +99,8 @@ namespace OpenCensus.Resource
         /// always be quoted.
         /// </summary>
         /// <param name="rawEnvironmentTags">Environment tags as a raw, comma separated string</param>
-        /// <returns>Environment Tags as a list</returns>
-        private static ITag[] ParseResourceLabels(string rawEnvironmentTags)
+        /// <returns>Environment Tags as a list.</returns>
+        internal static ITag[] ParseResourceLabels(string rawEnvironmentTags)
         {
             if (rawEnvironmentTags == null)
             {
@@ -133,13 +109,13 @@ namespace OpenCensus.Resource
             else
             {
                 var labels = new List<ITag>();
-                string[] rawLabels = rawEnvironmentTags.Split(LABEL_LIST_SPLITTER);
+                string[] rawLabels = rawEnvironmentTags.Split(LabelListSplitter);
 
                 Regex regex = new Regex("^\"|\"$", RegexOptions.Compiled);
                 
                 foreach (var rawLabel in rawLabels)
                 {
-                    string[] keyValuePair = rawLabel.Split(LABEL_KEY_VALUE_SPLITTER);
+                    string[] keyValuePair = rawLabel.Split(LabelKeyValueSplitter);
                     if (keyValuePair.Length != 2)
                     {
                         continue;
@@ -148,28 +124,37 @@ namespace OpenCensus.Resource
                     string key = keyValuePair[0].Trim();
                     string value = Regex.Replace(keyValuePair[1].Trim(), "^\"|\"$", string.Empty);
 
-                    Arguments.Check(IsValidAndNotEmpty(key), "Label key" + ERROR_MESSAGE_INVALID_CHARS);
-                    Arguments.Check(IsValid(value), "Label value" + ERROR_MESSAGE_INVALID_CHARS);
+                    if (!IsValidAndNotEmpty(key))
+                    {
+                        Log.InvalidCharactersInElement("Label key");
+                        return new ITag[0] { };
+                    }
 
-                    labels.Add(new Tag(new TagKey(key), new TagValue(value)));
+                    if (!IsValid(value))
+                    {
+                        Log.InvalidCharactersInElement("Label key");
+                        return new ITag[0] { };
+                    }
+
+                    labels.Add(new Tag(TagKey.Create(key), TagValue.Create(value)));
                 }
 
                 return labels.ToArray();
             }
         }
 
-        private static bool TryParseResourceType(string rawEnvironmentType, out string resourceType)
+        internal static bool TryParseResourceType(string rawEnvironmentType, out string resourceType)
         {
             if (string.IsNullOrEmpty(rawEnvironmentType))
             {
-                resourceType = GlobalResourceType;
+                resourceType = Constants.GlobalResourceType;
                 return false;
             }
 
-            if (rawEnvironmentType.Length > MaxResourceTypeNameLength)
+            if (rawEnvironmentType.Length > Constants.MaxResourceTypeNameLength)
             {
-                // TODO Log error $"Type {ERROR_MESSAGE_INVALID_CHARS}";
-                resourceType = GlobalResourceType;
+                Log.InvalidCharactersInElement(rawEnvironmentType);
+                resourceType = Constants.GlobalResourceType;
                 return false;
             }
 
@@ -179,18 +164,18 @@ namespace OpenCensus.Resource
 
         /// <summary>
         /// Checks whether given string is a valid printable ASCII string with a length not exeeding
-        /// <see cref="MaxResourceTypeNameLength"/> characters.
+        /// <see cref="Constants.MaxResourceTypeNameLength"/> characters.
         /// </summary>
         /// <param name="name">The string.</param>
         /// <returns>Whether given string is valid.</returns>
         private static bool IsValid(string name)
         {
-            return name.Length <= MaxResourceTypeNameLength && StringUtil.IsPrintableString(name);
+            return name.Length <= Constants.MaxResourceTypeNameLength && StringUtil.IsPrintableString(name);
         }
 
         /// <summary>
         /// Checks whether given string is a valid printable ASCII string with a length
-        /// greater than 0 and not exceeding <see cref="MaxLength"/> characters.
+        /// greater than 0 and not exceeding <see cref="Constants.MaxResourceTypeNameLength"/> characters.
         /// </summary>
         /// <param name="name">The string.</param>
         /// <returns>Whether given string is valid.</returns>